Fix Linux build

+++ wpilibc/src/main/native/include/frc/estimator/KalmanFilterLatencyCompensator.h
@@ -79,7 +79,7 @@ class KalmanFilterLatencyCompensator {
     // equal to the vision pose timestamp. We will now find the entry which is
     // closest in time to the requested timestamp.
 
-    int indexOfClosestEntry =
+    size_t indexOfClosestEntry =
         units::math::abs(timestamp - m_pastObserverSnapshots[index - 1].first) <
                 units::math::abs(timestamp -
                                  m_pastObserverSnapshots[index].first)
@@ -94,7 +94,7 @@ class KalmanFilterLatencyCompensator {
     // and apply correction based on the measurement. Then, we will go back
     // through all observer states until the present and apply past inputs to
     // get the present estimated state.
-    for (int i = indexOfClosestEntry; i < m_pastObserverSnapshots.size(); ++i) {
+    for (size_t i = indexOfClosestEntry; i < m_pastObserverSnapshots.size(); ++i) {
       auto& [key, snapshot] = m_pastObserverSnapshots[i];
 
       if (i == indexOfClosestEntry) {
